PVS V784: fix size of DR7 register
Even on 64bit CPUs it's a 32bit register.

Reviewed-on: https://review.haiku-os.org/c/1625

+++ headers/private/kernel/arch/x86/arch_user_debugger.h
@@ -101,7 +101,7 @@ struct arch_breakpoint {
 struct arch_team_debug_info {
 	struct arch_breakpoint	breakpoints[X86_BREAKPOINT_COUNT];
 
-	size_t					dr7;	// debug control register DR7
+	uint32_t				dr7;	// debug control register DR7
 };
 
 struct arch_thread_debug_info {
